extract plugin interfaces into corresponding plugin files
